# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "pattern": "^[A-Z][A-Z0-9_]*$",
                "type": "string"
              },
              "issues": {
                "items": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "enum": [
                        "INVALID_TYPE",
                        "INVALID_FORMAT",
                        "OUT_OF_RANGE",
                        "UNKNOWN_FIELD",
                        "INVALID_VALUE"
                      ],
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "source": {
                      "enum": [
                        "body",
                        "query",
                        "params"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "source",
                    "path",
                    "code"
                  ],
                  "type": "object"
                },
                "type": "array"
              },
              "metadata": {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "number"
                    },
                    {
                      "type": "boolean"
                    },
                    {
                      "enum": [
                        null
                      ],
                      "nullable": true,
                      "type": "string"
                    }
                  ]
                },
                "type": "object"
              },
              "status": {
                "maximum": 599,
                "minimum": 400,
                "type": "integer"
              },
              "traceId": {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              }
            },
            "required": [
              "code",
              "status",
              "traceId",
              "issues",
              "metadata"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "CreatedInvitationResponse": {
        "additionalProperties": false,
        "properties": {
          "acceptanceToken": {
            "minLength": 32,
            "type": "string"
          },
          "email": {
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d\\.\\d{3}(?:Z))$",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "administrator",
              "member",
              "viewer"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "accepted",
              "cancelled",
              "expired"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "status",
          "expiresAt",
          "acceptanceToken"
        ],
        "type": "object"
      },
...
      "OrganizationInvitationListResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "email": {
                  "format": "email",
                  "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
                  "type": "string"
                },
                "expiresAt": {
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d\\.\\d{3}(?:Z))$",
                  "type": "string"
                },
                "id": {
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "type": "string"
                },
                "role": {
                  "enum": [
                    "owner",
                    "administrator",
                    "member",
                    "viewer"
                  ],
                  "type": "string"
                },
                "status": {
                  "enum": [
                    "pending",
                    "accepted",
                    "cancelled",
                    "expired"
                  ],
                  "type": "string"
                }
              },
              "required": [
                "id",
                "email",
                "role",
                "status",
                "expiresAt"
              ],
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "OrganizationMemberListResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "displayName": {
                  "minLength": 1,
                  "nullable": true,
                  "type": "string"
                },
                "email": {
                  "format": "email",
                  "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
                  "type": "string"
                },
                "membershipId": {
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "type": "string"
                },
                "role": {
                  "enum": [
                    "owner",
                    "administrator",
                    "member",
                    "viewer"
                  ],
                  "type": "string"
                },
                "status": {
                  "enum": [
                    "active",
                    "suspended",
                    "removed"
                  ],
                  "type": "string"
                },
                "userId": {
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "type": "string"
                }
              },
              "required": [
                "membershipId",
                "userId",
                "email",
                "displayName",
                "role",
                "status"
              ],
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "Ardenfold API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations/current/invitations": {
      "get": {
        "operationId": "listOrganizationInvitations",
        "parameters": [
          {
            "in": "header",
            "name": "x-ardenfold-organization-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationInvitationListResponse"
                }
              }
            },
            "description": "",
            "headers": {
              "x-request-id": {
                "description": "Request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Machine-readable error; clients localize the error code.",
            "headers": {
              "x-request-id": {
                "description": "Request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "organizations"
        ]
      },
      "post": {
        "operationId": "createOrganizationInvitation",
        "parameters": [
          {
            "in": "header",
            "name": "x-ardenfold-organization-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedInvitationResponse"
                }
              }
            },
            "description": "",
            "headers": {
              "x-request-id": {
                "description": "Request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Machine-readable error; clients localize the error code.",
            "headers": {
              "x-request-id": {
                "description": "Request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "organizations"
        ]
      }
    },
    "/api/v1/organizations/current/invitations/{invitationId}": {
      "delete": {
        "operationId": "OrganizationInvitationsController_cancel",
        "parameters": [
          {
            "in": "path",
            "name": "invitationId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-ardenfold-organization-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "headers": {
              "x-request-id": {
                "description": "Request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Machine-readable error; clients localize the error code.",
            "headers": {
              "x-request-id": {
                "description": "Request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizations"
        ]
      }
    },
    "/api/v1/organizations/current/members": {
      "get": {
        "operationId": "listOrganizationMembers",
        "parameters": [
          {
            "in": "header",
            "name": "x-ardenfold-organization-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationMemberListResponse"
                }
              }
            },
            "description": "",
            "headers": {
              "x-request-id": {
                "description": "Request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Machine-readable error; clients localize the error code.",
            "headers": {
              "x-request-id": {
                "description": "Request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "organizations"
        ]
      }
    },
    "/api/v1/organizations/current/members/{membershipId}": {
      "delete": {
        "operationId": "OrganizationMembershipsController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "membershipId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-ardenfold-organization-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "headers": {
              "x-request-id": {
                "description": "Request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Machine-readable error; clients localize the error code.",
            "headers": {
              "x-request-id": {
                "description": "Request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizations"
        ]
      }
    },
    "/api/v1/organizations/current/members/{membershipId}/role": {
      "patch": {
        "operationId": "OrganizationMembershipsController_updateRole",
        "parameters": [
          {
            "in": "path",
            "name": "membershipId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-ardenfold-organization-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "headers": {
              "x-request-id": {
                "description": "Request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Machine-readable error; clients localize the error code.",
            "headers": {
              "x-request-id": {
                "description": "Request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizations"
        ]
      }
    },
    "/api/v1/organizations/current/members/{membershipId}/suspend": {
      "post": {
        "operationId": "OrganizationMembershipsController_suspend",
        "parameters": [
          {
            "in": "path",
            "name": "membershipId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-ardenfold-organization-id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "headers": {
              "x-request-id": {
                "description": "Request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Machine-readable error; clients localize the error code.",
            "headers": {
              "x-request-id": {
                "description": "Request correlation identifier.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "organizations"
        ]
      }
    },
...
  },
  "servers": [],
...
}